requirements.txt
apache-airflow-providers-odbc
pyodbc 
pandas 
numpy
sqlalchemy
mlxtend
tqdm